# Use script module for emoji-loader with settings exported via JSON