--- a/src/js/_enqueues/lib/emoji-loader.js
+++ b/src/js/_enqueues/lib/emoji-loader.js
@@ -359,7 +359,9 @@ settings.supports = {
 	everythingExceptFlag: true
 };
 
-// Create a promise for DOMContentLoaded since the worker logic may finish after the event has fired.
+// Since this is part of an async module, the emoji test worker script can start running before the document has fully
+// loaded and yet it may finish executing after the DOMContentLoaded event. So this promise is created here to ensure
+// the readyCallback is always called at or after DCL.
 const domReadyPromise = new Promise( ( resolve ) => {
 	document.addEventListener( 'DOMContentLoaded', resolve, {
 		once: true