--- a/src/wp-includes/formatting.php
+++ b/src/wp-includes/formatting.php
@@ -5985,8 +5985,10 @@ function _print_emoji_detection_script() {
 		)
 	);
 
+	$emoji_loader_script_path = '/js/wp-emoji-loader' . wp_scripts_get_suffix() . '.js';
 	wp_print_inline_script_tag(
-		file_get_contents( ABSPATH . WPINC . '/js/wp-emoji-loader' . wp_scripts_get_suffix() . '.js' ),
+		rtrim( file_get_contents( ABSPATH . WPINC . $emoji_loader_script_path ) ) . "\n" .
+		'//# sourceURL=' . includes_url( $emoji_loader_script_path ),
 		array(
 			'type' => 'module',
 		)