--- a/src/wp-includes/formatting.php
+++ b/src/wp-includes/formatting.php
@@ -5990,7 +5990,8 @@ function _print_emoji_detection_script() {
 		rtrim( file_get_contents( ABSPATH . WPINC . $emoji_loader_script_path ) ) . "\n" .
 		'//# sourceURL=' . includes_url( $emoji_loader_script_path ),
 		array(
-			'type' => 'module',
+			'type'  => 'module',
+			'async' => true,
 		)
 	);
 }