--- a/tests/phpunit/tests/formatting/emoji.php
+++ b/tests/phpunit/tests/formatting/emoji.php
@@ -7,7 +7,44 @@
 class Tests_Formatting_Emoji extends WP_UnitTestCase {
 
 	private $png_cdn = 'https://s.w.org/images/core/emoji/16.0.1/72x72/';
-	private $svn_cdn = 'https://s.w.org/images/core/emoji/16.0.1/svg/';
+	private $svg_cdn = 'https://s.w.org/images/core/emoji/16.0.1/svg/';
+
+	/**
+	 * @ticket 63842
+	 *
+	 * @covers ::_print_emoji_detection_script
+	 */
+	public function test_script_tag_printing() {
+		// `_print_emoji_detection_script()` assumes `wp-includes/js/wp-emoji-loader.js` is present:
+		self::touch( ABSPATH . WPINC . '/js/wp-emoji-loader.js' );
+		$output = get_echo( '_print_emoji_detection_script' );
+
+		$processor = new WP_HTML_Tag_Processor( $output );
+		$this->assertTrue( $processor->next_tag() );
+		$this->assertSame( 'SCRIPT', $processor->get_tag() );
+		$this->assertSame( 'wp-emoji-settings', $processor->get_attribute( 'id' ) );
+		$this->assertSame( 'application/json', $processor->get_attribute( 'type' ) );
+		$text     = $processor->get_modifiable_text();
+		$settings = json_decode( $text, true );
+		$this->assertIsArray( $settings );
+
+		$this->assertEqualSets(
+			array( 'baseUrl', 'ext', 'svgUrl', 'svgExt', 'source' ),
+			array_keys( $settings )
+		);
+		$this->assertSame( $this->png_cdn, $settings['baseUrl'] );
+		$this->assertSame( '.png', $settings['ext'] );
+		$this->assertSame( $this->svg_cdn, $settings['svgUrl'] );
+		$this->assertSame( '.svg', $settings['svgExt'] );
+		$this->assertIsArray( $settings['source'] );
+		$this->assertArrayHasKey( 'wpemoji', $settings['source'] );
+		$this->assertArrayHasKey( 'twemoji', $settings['source'] );
+		$this->assertTrue( $processor->next_tag() );
+		$this->assertSame( 'SCRIPT', $processor->get_tag() );
+		$this->assertSame( 'module', $processor->get_attribute( 'type' ) );
+		$this->assertNull( $processor->get_attribute( 'src' ) );
+		$this->assertFalse( $processor->next_tag() );
+	}
 
 	/**
 	 * @ticket 36525
@@ -20,7 +57,7 @@ class Tests_Formatting_Emoji extends WP_UnitTestCase {
 		$output = get_echo( '_print_emoji_detection_script' );
 
 		$this->assertStringContainsString( wp_json_encode( $this->png_cdn, JSON_HEX_TAG | JSON_UNESCAPED_SLASHES ), $output );
-		$this->assertStringContainsString( wp_json_encode( $this->svn_cdn, JSON_HEX_TAG | JSON_UNESCAPED_SLASHES ), $output );
+		$this->assertStringContainsString( wp_json_encode( $this->svg_cdn, JSON_HEX_TAG | JSON_UNESCAPED_SLASHES ), $output );
 	}
 
 	public function _filtered_emoji_svn_cdn( $cdn = '' ) {
@@ -42,7 +79,7 @@ class Tests_Formatting_Emoji extends WP_UnitTestCase {
 		$output = get_echo( '_print_emoji_detection_script' );
 
 		$this->assertStringContainsString( wp_json_encode( $this->png_cdn, JSON_HEX_TAG | JSON_UNESCAPED_SLASHES ), $output );
-		$this->assertStringNotContainsString( wp_json_encode( $this->svn_cdn, JSON_HEX_TAG | JSON_UNESCAPED_SLASHES ), $output );
+		$this->assertStringNotContainsString( wp_json_encode( $this->svg_cdn, JSON_HEX_TAG | JSON_UNESCAPED_SLASHES ), $output );
 		$this->assertStringContainsString( wp_json_encode( $filtered_svn_cdn, JSON_HEX_TAG | JSON_UNESCAPED_SLASHES ), $output );
 
 		remove_filter( 'emoji_svg_url', array( $this, '_filtered_emoji_svn_cdn' ) );
@@ -68,7 +105,7 @@ class Tests_Formatting_Emoji extends WP_UnitTestCase {
 
 		$this->assertStringContainsString( wp_json_encode( $filtered_png_cdn, JSON_HEX_TAG | JSON_UNESCAPED_SLASHES ), $output );
 		$this->assertStringNotContainsString( wp_json_encode( $this->png_cdn, JSON_HEX_TAG | JSON_UNESCAPED_SLASHES ), $output );
-		$this->assertStringContainsString( wp_json_encode( $this->svn_cdn, JSON_HEX_TAG | JSON_UNESCAPED_SLASHES ), $output );
+		$this->assertStringContainsString( wp_json_encode( $this->svg_cdn, JSON_HEX_TAG | JSON_UNESCAPED_SLASHES ), $output );
 
 		remove_filter( 'emoji_url', array( $this, '_filtered_emoji_png_cdn' ) );
 	}