<!--
Hi there! Thanks for contributing to WordPress!

Pull Requests in this GitHub repository **must** be linked to a ticket in the WordPress Core Trac instance (https://core.trac.wordpress.org), and are only used for code review. **No pull requests will be merged on GitHub.**

See the WordPress Handbook page on using PRs for Code Review more information: https://make.wordpress.org/core/handbook/contribute/git/github-pull-requests-for-code-review/

If this is your first time contributing, you may also find reviewing these guides first to be helpful:
- FAQs for New Contributors: https://make.wordpress.org/core/handbook/tutorials/faq-for-new-contributors/
- Contributing with Code Guide: https://make.wordpress.org/core/handbook/contribute/
- WordPress Coding Standards: https://make.wordpress.org/core/handbook/best-practices/coding-standards/
- Inline Documentation Standards: https://make.wordpress.org/core/handbook/best-practices/inline-documentation-standards/
- Browser Support Policies: https://make.wordpress.org/core/handbook/best-practices/browser-support/
- Proper spelling and grammar related best practices: https://make.wordpress.org/core/handbook/best-practices/spelling/
-->

<!-- Insert a description of your changes here -->

Trac ticket: https://core.trac.wordpress.org/ticket/63842

This splits up the inline script output in `_print_emoji_detection_script()` into two:

1. A `script` of type `application/json` which contains the `_wpemojiSettings` data.
2. A `script` of type `module` which contains `wp-emoji-loader.js`, which parses `_wpemojiSettings` out of the JSON. 

The result is this inline script is eliminated from blocking the HTML parser from processing the page (and rendering the page) while waiting for the JavaScript to execute. The `wp-emoji-loader.js` script does not need to run in the `head` because it does not actually proceed with loading emoji (if needed) until `DOMContentLoaded`.

I used the [`benchmark-web-vitals`](https://github.com/GoogleChromeLabs/wpp-research/tree/main/cli#benchmark-web-vitals) command from [GoogleChromeLabs/wpp-research](https://github.com/GoogleChromeLabs/wpp-research) to analyze the performance impact of this on a vanilla WordPress install on the Sample Page. For example:

```bash
npm run research -- benchmark-web-vitals --url http://localhost:8000/sample-page/  --output=csv --number=100
```

On a high-end machine (e.g. MacBook Pro with M4 Pro chip), the difference is negligible:

URL | Before | After | Diff (ms) | Diff (%)
-- | -- | -- | -- | --
FCP (median) | 94.75 | 94.3 | -0.45 | -0.47%
LCP (median) | 102.2 | 101.8 | -0.4 | -0.39%
TTFB (median) | 58.1 | 58 | -0.1 | -0.17%
LCP-TTFB (median) | 43.5 | 43.7 | 0.2 | 0.46%

However, when the CPU is throttled (via `--throttle-cpu`) to emulate a low-tier mobile phone, then there is a clear impact. I used CPU throttle factor of 20 because this value is close to how Chrome DevTools calibrates my CPU to emulate such a device:

<img width="276" height="179" alt="image" src="https://github.com/user-attachments/assets/a9e41f6a-ebdf-491d-b21d-59a4258ba711" />

The command I run on `trunk` and again on this branch:

```bash
npm run research -- benchmark-web-vitals --url http://localhost:8000/sample-page/  --output=csv --number=100 --throttle-cpu=20 
```

The results show a >5% improvement to LCP:

URL | Before | After | Diff (ms) | Diff (%)
-- | -- | -- | -- | --
FCP (median) | 322.6 | 301.35 | -21.25 | -6.59%
LCP (median) | 411.5 | 388.35 | -23.15 | -5.63%
TTFB (median) | 58.85 | 59.35 | 0.5 | 0.85%
LCP-TTFB (median) | 353.55 | 329.85 | -23.7 | -6.70%

This is the diff or a rendered page with Prettier formatting applied and `SCRIPT_DEBUG` enabled:

```diff
--- before.html	2025-08-19 12:46:30
+++ after.html	2025-08-19 12:47:16
@@ -28,20 +28,19 @@
 			title="WordPress Develop &raquo; Sample Page Comments Feed"
 			href="http://localhost:8000/sample-page/feed/"
 		/>
-		<script>
-			window._wpemojiSettings = {
-				baseUrl:
-					"https:\/\/s.w.org\/images\/core\/emoji\/16.0.1\/72x72\/",
-				ext: ".png",
-				svgUrl: "https:\/\/s.w.org\/images\/core\/emoji\/16.0.1\/svg\/",
-				svgExt: ".svg",
-				source: {
-					wpemoji:
-						"http:\/\/localhost:8000\/wp-includes\/js\/wp-emoji.js?ver=6.9-alpha-60093-src",
-					twemoji:
-						"http:\/\/localhost:8000\/wp-includes\/js\/twemoji.js?ver=6.9-alpha-60093-src",
-				},
-			};
+		<script id="wp-emoji-settings" type="application/json">
+			{
+				"baseUrl": "https://s.w.org/images/core/emoji/16.0.1/72x72/",
+				"ext": ".png",
+				"svgUrl": "https://s.w.org/images/core/emoji/16.0.1/svg/",
+				"svgExt": ".svg",
+				"source": {
+					"wpemoji": "http://localhost:8000/wp-includes/js/wp-emoji.js?ver=6.9-alpha-60093-src",
+					"twemoji": "http://localhost:8000/wp-includes/js/twemoji.js?ver=6.9-alpha-60093-src"
+				}
+			}
+		</script>
+		<script type="module">
 			/**
 			 * @output wp-includes/js/wp-emoji-loader.js
 			 */
@@ -58,6 +57,13 @@
 			 * @property {?Function} readyCallback
 			 */
 
+			// For compatibility with other scripts that read from this global.
+			window._wpemojiSettings = /** @type {WPEmojiSettings} */ (
+				JSON.parse(
+					document.getElementById("wp-emoji-settings").textContent,
+				)
+			);
+
 			/**
 			 * Support tests.
 			 * @typedef SupportTests
```


---
**This Pull Request is for code review only. Please keep all other discussion in the Trac ticket. Do not merge this Pull Request. See [GitHub Pull Requests for Code Review](https://make.wordpress.org/core/handbook/contribute/git/github-pull-requests-for-code-review/) in the Core Handbook for more details.**