@peterwilsoncc 2025-08-21
I've noticed a back-compat issue for scripts that depend on the global settings object.

```php
add_action( 'wp_enqueue_scripts', function() {
	// Script that uses _wpemojiSettings. Something GDPR maybe?
	wp_register_script( 'pwcc-no-deps-head', false, [], '1.0' );
	wp_enqueue_script( 'pwcc-no-deps-head' );
	wp_add_inline_script( 'pwcc-no-deps-head', 'console.log( "pwcc-no-deps-head", window._wpemojiSettings  );' );

	wp_register_script( 'pwcc-no-deps-foot', false, [], '1.0', true );
	wp_enqueue_script( 'pwcc-no-deps-foot' );
	wp_add_inline_script( 'pwcc-no-deps-foot', 'console.log( "pwcc-no-deps-foot", window._wpemojiSettings  );' );
} );
```

On `trunk` both of these scripts log the settings object, on this branch they both log `undefined` .

Searching the plugin repo for [_wpemojiSettings](https://wpdirectory.net/search/01K35DZ5Q7HWX1SNYH2R9DHVAP) shows that it's referenced in WooCommerce Payments, 700K active sites in the file registered by this [block of code](https://github.com/Automattic/woocommerce-payments/blob/2735cf15ea7f088f092b452162e83de79d822541/includes/multi-currency/CurrencySwitcherBlock.php#L63-L69), so I think we'll need some care if this is to change.

A few of the other results appear to be false positives.