@westonruter 2025-08-21
Here's where `_wpemojiSettings` is referenced in the plugin:

https://github.com/Automattic/woocommerce-payments/blob/7d119265c70a452def2eccc11ef4f1729f904f16/includes/multi-currency/client/blocks/currency-switcher.js#L143

Specifically it's I'm a block's edit function in the editor, which would surely be executed after `DOMContentLoaded`, so it doesn't seem like it would be a problem.